requirements.txt
# general modules
django==4.2.20
djangorestframework==3.15.2
python-decouple==3.8
Pillow==10.4.0
coreapi
coreschema

# Third party modules
markdown==3.4.3
django-filter==23.2
drf-yasg[validation]==1.21.11
djangorestframework-simplejwt==5.3.1
djoser==2.2.1
# deployment module

# email third party module
django-mail-templated==2.6.5



# refactoring and tests
black==25.11.0
flake8==7.3.0
pytest==9.0.1
pytest-django==4.11.1